refactor: move all cli commands into its own cli folder
