+++ .github/workflows/ui.yml
@@ -109,7 +109,7 @@ jobs:
         working-directory: application/backend
         run: |
           export PYTHONPATH=.
-          uv run src/cli.py gen-api --target-path openapi-spec.json
+          uv run -m src.cli gen-api --target-path openapi-spec.json
 
       - uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
         with:
