+++ application/backend/run.sh
@@ -6,15 +6,11 @@ set -euo pipefail
 #
 # Features:
 # - Runs database migrations on every start (idempotent via Alembic)
-# - Optionally seeds the database before starting the server by setting:
-#     SEED_DB=true
 #
 # Usage:
-#   SEED_DB=true ./run.sh       # Seed database before launching server
-#   ./run.sh                    # Run server without seeding
+#   ./run.sh                    # Run server
 #
 # Environment variables:
-#   SEED_DB       If set to "true", runs database seeding before starting.
 #   APP_MODULE    Python module to run (default: src/main.py)
 #   UV_CMD        Command to launch Uvicorn (default: "uv run")
 #
@@ -23,7 +19,6 @@ set -euo pipefail
 # - Python modules and dependencies installed correctly
 # -----------------------------------------------------------------------------
 
-SEED_DB=${SEED_DB:-false}
 APP_MODULE=${APP_MODULE:-src/main.py}
 UV_CMD=${UV_CMD:-uv run --no-sync}
 
@@ -34,15 +29,8 @@ export PYTHONPATH=.
 # already-applied migrations. This ensures the persistent volume
 # has an up-to-date schema regardless of how it was created.
 echo "Running database migrations..."
-$UV_CMD src/cli.py migrate
-
-if [[ "$SEED_DB" == "true" ]]; then
-	echo "Seeding the database..."
-	$UV_CMD application/cli.py init-db
-	$UV_CMD application/cli.py seed --with-model=True
-fi
+$UV_CMD -m src.cli db migrate
 
 echo "Starting FastAPI server..."
-
 echo $UV_CMD "$APP_MODULE"
 exec $UV_CMD "$APP_MODULE"
