+++ application/backend/src/cli/__main__.py
@@ -1,0 +1,7 @@
+"""Module entrypoint for `python -m src.cli`."""
+
+from . import cli
+
+
+if __name__ == "__main__":
+    cli()
